# Aurora-Motor-Engine/hybrid-agent-fabric/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aurora Motor Engine API
  description: |
    Aurora Motor Engine — Hybrid Agent Fabric Control API.
    
    Bu API, Aurora bilişsel motorunun tüm bileşenlerini yönetir:
    - Session yönetimi ve sohbet
    - Aurora hafıza grafı
    - Dünya modeli
    - Proaktif initiative motoru
    - Evrim ve skill yönetimi
    - Bilişsel orkestratör
    - Topluluk (Society) sistemi
    - Platform entegrasyonları (Telegram, Discord, Slack, vb.)
    - MCP server yönetimi
    - Öğrenme ve iyileştirme
    - Otomasyon ve zamanlama
  version: 1.65.0
  contact:
    name: Aurora Motor Engine
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /v1/cognitive-state/tasks/{taskId}:
    get:
      tags: [cognitive-state]
      operationId: getV1CognitivestateTasksTaskId
      x-source: apps/control-api/src/routes/cognitive.ts
      x-schema-status: unspecified
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
...
  /v1/event-bus/trace/{traceId}:
    get:
      tags: [event-bus]
      operationId: getV1EventbusTraceTraceId
      x-source: apps/control-api/src/routes/cognitive.ts
      x-schema-status: unspecified
      parameters:
        - name: traceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
...
  /v1/meta-controller/alerts/{alertId}/acknowledge:
    post:
      tags: [meta-controller]
      operationId: postV1MetacontrollerAlertsAlertIdAcknowledge
      x-source: apps/control-api/src/routes/meta-controller.ts
      x-schema-status: unspecified
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: haf_session
...